requirements.txt
openai-whisper
transformers
torch
scikit-learn
pandas
numpy
streamlit
flask
vaderSentiment
spacy
pydub